requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyperclip==1.9.0
pywin32-ctypes==0.2.3
setuptools==75.8.2
